WAKDEBOT/requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
numpy
streamlit==1.32.0
groq==0.4.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1 